+++ threads.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   threads.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sukwon <sukwon@student.hive.fi>            +#+  +:+       +#+        */
+/*   By: skwon2 <skwon2@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 23:05:41 by sukwon            #+#    #+#             */
-/*   Updated: 2024/07/04 12:07:40 by sukwon           ###   ########.fr       */
+/*   Updated: 2024/07/05 16:36:45 by skwon2           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int detach_thread_philos(t_data *data)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < data->num_philos)
 	{
 		if (pthread_detach(data->philos[i].philo) != 0)
 		{
-			printf("data->philos[%d].philo : pthread_detach has been failed\n", i);
+			printf("data->philos[%ld].philo : pthread_detach has been failed\n", i);
 			return (EXIT_FAILURE);
 		}
 		i++;
@@ -31,7 +31,7 @@ int detach_thread_philos(t_data *data)
 
 int create_thread_philos(t_data *data)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 
@@ -44,7 +44,7 @@ int create_thread_philos(t_data *data)
 	{
 		if (pthread_create(&(data->philos[i].philo), NULL, philo_routine, &data->philos[i]) != 0)
 		{
-			printf("data->philos[%d].philo : pthread_create has been failed\n", i);
+			printf("data->philos[%ld].philo : pthread_create has been failed\n", i);
 			if (detach_thread_philos(data) == 1)
 				return (EXIT_FAILURE);
 		}
@@ -55,7 +55,7 @@ int create_thread_philos(t_data *data)
 
 int wait_threads(t_data *data)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (pthread_join(data->monitor, NULL) != 0) // NULL can be some varialbes that if the threads has a return value to store
@@ -67,7 +67,7 @@ int wait_threads(t_data *data)
 	{
 		if (pthread_join(data->philos[i].philo, NULL) != 0) // NULL can be some varialbes that if the threads has a return value to store
 		{
-			printf("data->philos[%d].philo : pthread_join has been failed\n", i);
+			printf("data->philos[%ld].philo : pthread_join has been failed\n", i);
 			if (detach_thread_philos(data) == 1)
 				return (EXIT_FAILURE);
 		}
